move SamModel into serviceTest/Sam, change function names back to original
